NO-ISSUE: Move `TmpFS` to the testing package
This patch moves the `TmpFS` function to the `testing` package, so that
it can be used in other tests.
