+++ ztp/internal/templating/suite_test.go
@@ -15,10 +15,6 @@ License.
 package templating
 
 import (
-	"errors"
-	"io/fs"
-	"os"
-	"path/filepath"
 	"testing"
 
 	. "github.com/onsi/ginkgo/v2/dsl/core"
@@ -28,50 +24,4 @@ import (
 func TestTemplating(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Templating")
-}
-
-// TmpFS creates a temporary directory containing the given files, and then creates a fs.FS object
-// that can be used to access it.
-//
-// The files are specified as pairs of full path names and content. For example, to create a file
-// named `mydir/myfile.yaml` containig some YAML text and a file `yourdir/yourfile.json` containing
-// some JSON text:
-//
-//	dir, fsys = TmpFS(
-//		"mydir/myfile.yaml",
-//		`
-//			name: Joe
-//			age: 52
-//		`,
-//		"yourdir/yourfile.json",
-//		`{
-//			"name": "Mary",
-//			"age": 59
-//		}`
-//	)
-//
-// Directories are created automatically when they contain at least one file or subdirectory.
-//
-// The caller is responsible for removing the directory once it is no longer needed.
-func TmpFS(args ...string) (dir string, fsys fs.FS) {
-	Expect(len(args) % 2).To(BeZero())
-	dir, err := os.MkdirTemp("", "*.testfs")
-	Expect(err).ToNot(HaveOccurred())
-	for i := 0; i < len(args)/2; i++ {
-		name := args[2*i]
-		text := args[2*i+1]
-		file := filepath.Join(dir, name)
-		sub := filepath.Dir(file)
-		_, err = os.Stat(sub)
-		if errors.Is(err, os.ErrNotExist) {
-			err = os.MkdirAll(sub, 0700)
-			Expect(err).ToNot(HaveOccurred())
-		} else {
-			Expect(err).ToNot(HaveOccurred())
-		}
-		err = os.WriteFile(file, []byte(text), 0600)
-		Expect(err).ToNot(HaveOccurred())
-	}
-	fsys = os.DirFS(dir)
-	return
 }
