+++ ztp/internal/testing/environment_ca.go
@@ -12,7 +12,7 @@ implied. See the License for the specific language governing permissions and lim
 License.
 */
 
-package internal
+package testing
 
 import (
 	"crypto/rand"
